==> hotel_reviews_classification/1 data_prep.ipynb <==
# import dataset
import pandas as pd

df = pd.read_csv('data/tripadvisor_hotel_reviews.csv')
df

# split into train 80%, dev 10%, test 10%
from sklearn.model_selection import train_test_split

train, dev_test = train_test_split(df, test_size = 0.2, random_state = 333)
dev, test = train_test_split(dev_test, test_size = 0.5, random_state = 222)
 
train.to_csv('data/train.tsv', sep='\t', index=False)
dev.to_csv('data/dev.tsv', sep='\t', index=False)
test.to_csv('data/test.tsv', sep='\t', index=False)
